Milestone D3R3, continue to D4R1

+++ lawnchair/src/app/lawnchair/ui/preferences/destinations/HomeScreenPreferences.kt
@@ -40,7 +40,6 @@ import app.lawnchair.ui.preferences.components.controls.ClickablePreference
 import app.lawnchair.ui.preferences.components.controls.ListPreference
 import app.lawnchair.ui.preferences.components.controls.SliderPreference
 import app.lawnchair.ui.preferences.components.controls.SwitchPreference
-import app.lawnchair.ui.preferences.components.layout.ExpandAndShrink
 import app.lawnchair.ui.preferences.components.layout.PreferenceGroup
 import app.lawnchair.ui.preferences.components.layout.PreferenceLayout
 import app.lawnchair.ui.preferences.navigation.HomeScreenGrid
@@ -76,7 +75,10 @@ fun HomeScreenPreferences(
         PreferenceGroup(heading = stringResource(id = R.string.general_label)) {
             val addIconToHomeAdapter = prefs.addIconToHome.getAdapter()
             val isDeckLayoutAdapter = prefs2.deckLayout.getAdapter()
-            ExpandAndShrink(visible = !isDeckLayoutAdapter.state.value) {
+            Item(
+                "add_icon_to_home",
+                !isDeckLayoutAdapter.state.value,
+            ) {
                 SwitchPreference(
                     checked = (!lockHomeScreenAdapter.state.value && addIconToHomeAdapter.state.value) || isDeckLayoutAdapter.state.value,
                     onCheckedChange = addIconToHomeAdapter::onChange,
@@ -85,78 +87,113 @@ fun HomeScreenPreferences(
                     enabled = lockHomeScreenAdapter.state.value.not(),
                 )
             }
-            GestureHandlerPreference(
-                adapter = prefs2.doubleTapGestureHandler.getAdapter(),
-                label = stringResource(id = R.string.gesture_double_tap),
-            )
+            Item {
+                GestureHandlerPreference(
+                    adapter = prefs2.doubleTapGestureHandler.getAdapter(),
+                    label = stringResource(id = R.string.gesture_double_tap),
+                )
+            }
+            Item {
+                SwitchPreference(
+                    prefs.infiniteScrolling.getAdapter(),
+                    label = stringResource(id = R.string.infinite_scrolling_label),
+                    description = stringResource(id = R.string.infinite_scrolling_description),
+                )
+            }
         }
+        val feedAvailable = OverlayCallbackImpl.minusOneAvailable(LocalContext.current)
+        val enableFeedAdapter = prefs2.enableFeed.getAdapter()
         PreferenceGroup(heading = stringResource(id = R.string.minus_one)) {
-            val feedAvailable = OverlayCallbackImpl.minusOneAvailable(LocalContext.current)
-            val enableFeedAdapter = prefs2.enableFeed.getAdapter()
-            SwitchPreference(
-                adapter = enableFeedAdapter,
-                label = stringResource(id = R.string.minus_one_enable),
-                description = if (feedAvailable) null else stringResource(id = R.string.minus_one_unavailable),
-                enabled = feedAvailable,
-            )
-            ExpandAndShrink(visible = feedAvailable && enableFeedAdapter.state.value) {
+            Item {
+                SwitchPreference(
+                    adapter = enableFeedAdapter,
+                    label = stringResource(id = R.string.minus_one_enable),
+                    description = if (feedAvailable) null else stringResource(id = R.string.minus_one_unavailable),
+                    enabled = feedAvailable,
+                )
+            }
+            Item(
+                key = "feed_pref",
+                visible = feedAvailable && enableFeedAdapter.state.value,
+            ) {
                 FeedPreference()
             }
         }
         PreferenceGroup(heading = stringResource(R.string.style)) {
-            HomeScreenTextColorPreference()
-            OverlayHandlerPreference(
-                adapter = prefs2.closingAppOverlay.getAdapter(),
-                label = stringResource(id = R.string.app_closing_animation),
-            )
+            Item { HomeScreenTextColorPreference() }
+            Item {
+                OverlayHandlerPreference(
+                    adapter = prefs2.closingAppOverlay.getAdapter(),
+                    label = stringResource(id = R.string.app_closing_animation),
+                )
+            }
         }
         PreferenceGroup(heading = stringResource(id = R.string.wallpaper)) {
-            SwitchPreference(
-                prefs.wallpaperScrolling.getAdapter(),
-                label = stringResource(id = R.string.wallpaper_scrolling_label),
-            )
-            if (Utilities.ATLEAST_R) {
+            Item {
+                SwitchPreference(
+                    prefs.wallpaperScrolling.getAdapter(),
+                    label = stringResource(id = R.string.wallpaper_scrolling_label),
+                )
+            }
+            Item(
+                "wallpaper_depth_effect",
+                Utilities.ATLEAST_R,
+            ) {
                 SwitchPreference(
                     prefs2.wallpaperDepthEffect.getAdapter(),
                     label = stringResource(id = R.string.wallpaper_depth_effect_label),
                     description = stringResource(id = R.string.wallpaper_depth_effect_description),
                 )
             }
-            SwitchPreference(
-                adapter = prefs2.showTopShadow.getAdapter(),
-                label = stringResource(id = R.string.show_sys_ui_scrim),
-            )
+            Item {
+                SwitchPreference(
+                    adapter = prefs2.showTopShadow.getAdapter(),
+                    label = stringResource(id = R.string.show_sys_ui_scrim),
+                )
+            }
         }
+        val columns by prefs.workspaceColumns.getAdapter()
+        val rows by prefs.workspaceRows.getAdapter()
         PreferenceGroup(heading = stringResource(id = R.string.layout)) {
-            val columns by prefs.workspaceColumns.getAdapter()
-            val rows by prefs.workspaceRows.getAdapter()
-            NavigationActionPreference(
-                label = stringResource(id = R.string.home_screen_grid),
-                destination = HomeScreenGrid,
-                subtitle = stringResource(id = R.string.x_by_y, columns, rows),
-            )
-            SwitchPreference(
-                adapter = lockHomeScreenAdapter,
-                label = stringResource(id = R.string.home_screen_lock),
-                description = stringResource(id = R.string.home_screen_lock_description),
-            )
+            Item {
+                NavigationActionPreference(
+                    label = stringResource(id = R.string.home_screen_grid),
+                    destination = HomeScreenGrid,
+                    subtitle = stringResource(id = R.string.x_by_y, columns, rows),
+                )
+            }
+            Item {
+                SwitchPreference(
+                    adapter = lockHomeScreenAdapter,
+                    label = stringResource(id = R.string.home_screen_lock),
+                    description = stringResource(id = R.string.home_screen_lock_description),
+                )
+            }
         }
         PreferenceGroup(heading = stringResource(id = R.string.popup_menu)) {
-            LauncherPopupPreferenceItem()
+            Item { LauncherPopupPreferenceItem() }
         }
+        val showStatusBarAdapter = prefs2.showStatusBar.getAdapter()
         PreferenceGroup(heading = stringResource(id = R.string.status_bar_label)) {
-            val showStatusBarAdapter = prefs2.showStatusBar.getAdapter()
-            SwitchPreference(
-                adapter = showStatusBarAdapter,
-                label = stringResource(id = R.string.show_status_bar),
-            )
-            ExpandAndShrink(visible = showStatusBarAdapter.state.value) {
+            Item {
+                SwitchPreference(
+                    adapter = showStatusBarAdapter,
+                    label = stringResource(id = R.string.show_status_bar),
+                )
+            }
+            Item(
+                "dark_status_bar",
+                showStatusBarAdapter.state.value,
+            ) {
                 SwitchPreference(
                     adapter = prefs2.darkStatusBar.getAdapter(),
                     label = stringResource(id = R.string.dark_status_bar_label),
                 )
             }
-            ExpandAndShrink(visible = showStatusBarAdapter.state.value && LawnchairApp.isRecentsEnabled) {
+            Item(
+                "status_bar_clock",
+                showStatusBarAdapter.state.value && LawnchairApp.isRecentsEnabled,
+            ) {
                 SwitchPreference(
                     adapter = prefs2.statusBarClock.getAdapter(),
                     label = stringResource(id = R.string.status_bar_clock_label),
@@ -164,20 +201,27 @@ fun HomeScreenPreferences(
                 )
             }
         }
+        val homeScreenLabelsAdapter = prefs2.showIconLabelsOnHomeScreen.getAdapter()
         PreferenceGroup(heading = stringResource(id = R.string.icons)) {
-            SliderPreference(
-                label = stringResource(id = R.string.icon_sizes),
-                adapter = prefs2.homeIconSizeFactor.getAdapter(),
-                step = 0.1f,
-                valueRange = 0.5F..1.5F,
-                showAsPercentage = true,
-            )
-            val homeScreenLabelsAdapter = prefs2.showIconLabelsOnHomeScreen.getAdapter()
-            SwitchPreference(
-                adapter = homeScreenLabelsAdapter,
-                label = stringResource(id = R.string.show_labels),
-            )
-            ExpandAndShrink(visible = homeScreenLabelsAdapter.state.value) {
+            Item {
+                SliderPreference(
+                    label = stringResource(id = R.string.icon_sizes),
+                    adapter = prefs2.homeIconSizeFactor.getAdapter(),
+                    step = 0.1f,
+                    valueRange = 0.5F..1.5F,
+                    showAsPercentage = true,
+                )
+            }
+            Item {
+                SwitchPreference(
+                    adapter = homeScreenLabelsAdapter,
+                    label = stringResource(id = R.string.show_labels),
+                )
+            }
+            Item(
+                "workspace_label_size",
+                homeScreenLabelsAdapter.state.value,
+            ) {
                 SliderPreference(
                     label = stringResource(id = R.string.label_size),
                     adapter = prefs2.homeIconLabelSizeFactor.getAdapter(),
@@ -189,34 +233,44 @@ fun HomeScreenPreferences(
         }
         val overrideRepo = IconOverrideRepository.INSTANCE.get(LocalContext.current)
         val customIconsCount by remember { overrideRepo.observeCount() }.collectAsStateBlocking()
-        ExpandAndShrink(visible = customIconsCount > 0) {
+        if (customIconsCount > 0) {
             PreferenceGroup {
-                ClickablePreference(
-                    label = stringResource(id = R.string.reset_custom_icons),
-                    confirmationText = stringResource(id = R.string.reset_custom_icons_confirmation),
-                    onClick = { scope.launch { overrideRepo.deleteAll() } },
-                )
+                Item {
+                    ClickablePreference(
+                        label = stringResource(id = R.string.reset_custom_icons),
+                        confirmationText = stringResource(id = R.string.reset_custom_icons_confirmation),
+                        onClick = { scope.launch { overrideRepo.deleteAll() } },
+                    )
+                }
             }
         }
         PreferenceGroup(heading = stringResource(id = R.string.widget_button_text)) {
-            SwitchPreference(
-                adapter = prefs2.roundedWidgets.getAdapter(),
-                label = stringResource(id = R.string.force_rounded_widgets),
-            )
-            SwitchPreference(
-                adapter = prefs2.allowWidgetOverlap.getAdapter(),
-                label = stringResource(id = R.string.allow_widget_overlap),
-            )
-            SwitchPreference(
-                adapter = prefs2.widgetUnlimitedSize.getAdapter(),
-                label = stringResource(id = R.string.widget_unlimited_size_label),
-                description = stringResource(id = R.string.widget_unlimited_size_description),
-            )
-            SwitchPreference(
-                adapter = prefs2.forceWidgetResize.getAdapter(),
-                label = stringResource(id = R.string.force_widget_resize_label),
-                description = stringResource(id = R.string.force_widget_resize_description),
-            )
+            Item {
+                SwitchPreference(
+                    adapter = prefs2.roundedWidgets.getAdapter(),
+                    label = stringResource(id = R.string.force_rounded_widgets),
+                )
+            }
+            Item {
+                SwitchPreference(
+                    adapter = prefs2.allowWidgetOverlap.getAdapter(),
+                    label = stringResource(id = R.string.allow_widget_overlap),
+                )
+            }
+            Item {
+                SwitchPreference(
+                    adapter = prefs2.widgetUnlimitedSize.getAdapter(),
+                    label = stringResource(id = R.string.widget_unlimited_size_label),
+                    description = stringResource(id = R.string.widget_unlimited_size_description),
+                )
+            }
+            Item {
+                SwitchPreference(
+                    adapter = prefs2.forceWidgetResize.getAdapter(),
+                    label = stringResource(id = R.string.force_widget_resize_label),
+                    description = stringResource(id = R.string.force_widget_resize_description),
+                )
+            }
         }
     }
 }
